// frontend/src/types.ts
export interface AppContextType {
  drawerOpen: boolean;
  drawerRL: string;
  tPadSensitivity: number;
  mWheelSensitivity: number;
  buttonSW1: boolean;
  buttonSW4: boolean;
  button23: number;
  autoCollapse: boolean;
  mouseWheelMenuType: number;
  sidebarLayoutSettingMenu: boolean;
  sidebarMouseWheelMenu: boolean;
  sidebarKeyboardMenu: boolean;
  sidebarVolumeMenu: boolean;
  sidebarSettingMenu: boolean;
  sidebarThemeMenu: boolean;
  sidebarLanguageMenu: boolean;
  primaryColor: string;
  secondaryColor: string;
  openMenuSW: number;
  keyboardDataSendMod: string;
  i18n: string;
  setDrawerOpen: (value: boolean) => void;
  setDrawerRL: (value: string) => void;
  setTPadSensitivity: (value: number) => void;
  setMWheelSensitivity: (value: number) => void;
  setButtonSW1: (value: boolean) => void;
  setButtonSW4: (value: boolean) => void;
  setButton23: (value: number) => void;
  setAutoCollapse: (value: boolean) => void;
  setMouseWheelMenuType: (value: number) => void;
  setSidebarLayoutSettingMenu: (value: boolean) => void;
  setSidebarMouseWheelMenu: (value: boolean) => void;
  setSidebarKeyboardMenu: (value: boolean) => void;
  setSidebarVolumeMenu: (value: boolean) => void;
  setSidebarSettingMenu: (value: boolean) => void;
  setSidebarThemeMenu: (value: boolean) => void;
  setSidebarLanguageMenu: (value: boolean) => void;
  setPrimaryColor: (value: string) => void;
  setSecondaryColor: (value: string) => void;
  setOpenMenuSW: (value: number) => void;
  setKeyboardDataSendMod: (value: string) => void;
  setI18n: (value: string) => void;
  toggleButtonSW1: () => void;
  toggleButtonSW4: () => void;
  toggleAutoCollapse: () => void;
  toggleSidebarLayoutSettingMenu: () => void;
  toggleSidebarMouseWheelMenu: () => void;
  toggleSidebarKeyboardMenu: () => void;
  toggleSidebarVolumeMenu: () => void;
  toggleSidebarSettingMenu: () => void;
  toggleSidebarThemeMenu: () => void;
  toggleSidebarLanguageMenu: () => void;
}